allela_backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
supabase==2.4.6
stripe==9.9.0
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0
